requirements.txt
requests
python-dotenv
pydub
rich
xmltodict
youtube-search
pytubefix
mutagen
moviepy